test(app): separate real-model Nemotron glue for coverage hygiene
Patch coverage was diluted by code only exercisable with the real ~0.6 GB
CoreML models, which the default xctest lane can't run. Two moves, no logic
change:

- Move the real seam implementations (NemotronAsrManager,
  FluidVADBoundaryDetector, PartialCollector) out of
  NemotronStreamingCaptionSession.swift into NemotronAsrManager.swift, so the
  session's unit-tested logic isn't mixed with untestable model glue in one
  file. (Same pattern as LiveCaptionsOverlay vs the extracted, tested
  opacity math.)
- Codecov-ignore that file + LiveTranscriptionController+Nemotron.swift (the
  real-model pipeline builder) — both wrap FluidAudio's model + Silero VAD
  and are exercised end-to-end by the gated NemotronFinishDriftTests + live
  recording, with their mockable seams unit-tested. Mirrors the existing
  DebugTapFault / FluidDiarizer+SortformerEmbeddings ignores.

The original file's now-unused FluidAudio import is dropped (the session
drives the models only through the seam protocols).
